feat: save alert 창
